moves autozion to exercise4, splits it into multiple files and cleans the code up a bit
